Copy dotfiles including hidden ones

+++ 08-arch-install-dotfiles.sh
@@ -4,13 +4,19 @@
 CUR_USER=$(whoami)
 source /home/$CUR_USER/install-scripts/values.conf
 
+if ! [[ $EUID -ne 0 ]]; then
+	echo "Do not run this script as sudo/root."
+	exit 1
+fi
+
 # Clone and copy dot files
 git clone https://github.com/ethmth/arch-dots.git /home/$CUR_USER/arch-dots
+shopt -s dotglob
 cp -r /home/$CUR_USER/arch-dots/home/* /home/$CUR_USER/
 rm -rf /home/$CUR_USER/arch-dots
 
 
 # Install Magic Status Executables
 if (( PLASMA )); then
-    git clone https://github.com/ethmth/magic-status-executables.git /usr/share/plasma/plasmoids/com.github.ethmth.magic-status-executables
+    sudo -k git clone https://github.com/ethmth/magic-status-executables.git /usr/share/plasma/plasmoids/com.github.ethmth.magic-status-executables
 fi
